Add a h3 hexagon layer to the deckgl widget

Add a h3 hexagon layer to the deckgl widget

add_h3_hexagon_layer(deckgl, id = "h3-hexagon-layer", data = NULL,
  properties = list(), ...)

Arguments

deckgl
deckgl widget

id
id of the layer

data
url to fetch data from or data object

properties
named list of properties with names corresponding to the properties defined in the deckgl-api-reference for the given layer class, additionally there is a getTooltip property (callback) showing a tooltip when the mouse enters an object, e. g. getTooltip = JS("object => object.name")

...
more properties (will be added to the properties object), useful if you want to use a properties object for more than one layer

See also

Examples

## @knitr h3-hexagon-layer-layer
h3_cells <- system.file("sample-data/h3-cells.csv", package = "deckgl") %>%
  read.csv()

properties <- list(
  getHexagon = get_property("h3_index"),
  getFillColor =JS("d => [255, (1 - d.count / 500) * 255, 0]"),
  getElevation = get_property("count"),
  elevationScale = 20,
  getTooltip = JS("object => `${object.h3_index}: ${object.count}`")
)

deck <- deckgl(zoom = 11, pitch = 35) %>%
  add_h3_hexagon_layer(data = h3_cells, properties = properties) %>%
  add_mapbox_basemap()

if (interactive()) deck