Add a line layer to the deckgl widget

The LineLayer renders flat lines joining pairs of source and target points, specified as latitude/longitude coordinates.

add_line_layer(deckgl, id = "line-layer", data = NULL,
  properties = list(), ...)

Arguments

deckgl
deckgl widget

id
id of the layer

data
url to fetch data from or data object

properties
named list of properties with names corresponding to the properties defined in the deckgl-api-reference for the given layer class, additionally there is a getTooltip property (callback) showing a tooltip when the mouse enters an object, e. g. getTooltip = JS("object => object.name")

...
more properties (will be added to the properties object), useful if you want to use a properties object for more than one layer

See also

Examples

## @knitr line-layer
sample_data <- paste0(
  "https://raw.githubusercontent.com/",
  "uber-common/deck.gl-data/",
  "master/website/bart-segments.json"
)

properties <- list(
  pickable = TRUE,
  getStrokeWidth = 12,
  getSourcePosition = get_property("from.coordinates"),
  getTargetPosition = get_property("to.coordinates"),
  getColor = JS("d => [Math.sqrt(d.inbound + d.outbound), 140, 0]"),
  getTooltip = JS("object => `${object.from.name} to ${object.to.name}`")
)

deck <- deckgl(zoom = 10, pitch = 20) %>%
  add_line_layer(data = sample_data, properties = properties) %>%
  add_mapbox_basemap()

if (interactive()) deck