Add a base map from mapbox to the deckgl widget

Add a base map from mapbox to the deckgl widget

add_mapbox_basemap(deckgl, style = "mapbox://styles/mapbox/light-v9",
  token = Sys.getenv("MAPBOX_API_TOKEN"))

Arguments

deckgl
deckgl widget

style
map style

token
mapbox API access token

Value

deckgl widget